Moved db models to single file, app is WORKING, global variable working for user info
