+++ client/components/Login.jsx
@@ -10,14 +10,13 @@ const Login = (props) => {
     const [isSubmitted, setIsSubmitted] = useState(false);
     const [whoLogged, setWhoLogged] = useState()
     const userReset = props.setUser;
+    const userProfile = props.user;
 
     useEffect(()=>{
 
-    },[isSubmitted, whoLogged])
+    },[isSubmitted, whoLogged, userProfile])
 
-    const navigate = useNavigate();
-
-    const handleSubmit = async (e) => {
+    const handleSubmit = (e) => {
 
         e.preventDefault();
 
@@ -32,17 +31,17 @@ const Login = (props) => {
             url: 'http://localhost:3000/login', 
             data: { username, password}
         })
-         .then(async (res) => {
+         .then((res) => {
             console.log('Login.jsx .then res.data: ',res.data);
             console.log('logged in?: ', res.data.loggedIn); // true
             console.log('who logged in? ',res.data.user.username)
             //navigate(`${res.data.path}`, {replace:true})
             setIsSubmitted(res.data.loggedIn)
             setWhoLogged(res.data.user.username.toUpperCase());
-            await userReset(prev => {
+            userReset(prev => {
                 console.log(prev)
                 console.log(res.data.user)
-                return res.data.user.username.toUpperCase;
+                return res.data.user.username.toUpperCase();
             })
             console.log('Finished Logging in');
         });
@@ -60,7 +59,7 @@ const Login = (props) => {
 
     return(
         <div className="login-div">
-            {isSubmitted ? <div className="login-success"><h3>Successfully logged in!</h3><h3>Welcome {props.user}!</h3><Link to='/' id="enter-link">Click Here</Link></div> : renderForm}
+            {isSubmitted ? <div className="login-success"><h3>Successfully logged in!</h3><h3>Welcome {userProfile}!</h3><Link to='/' id="enter-link">Click Here</Link></div> : renderForm}
         </div>
     )
 }
