+++ server/dbModels/sessionModel.js
@@ -1,0 +1,10 @@
+//server/sessionModel.js
+// const mongoose = require('mongoose');
+// const Schema = mongoose.Schema;
+
+// const sessionSchema = new Schema({
+//     cookieId: { type: String, required: true, unique: true },
+//     createdAt: { type: Date, expires: 30, default: Date.now }
+//   });
+  
+//   module.exports = mongoose.model('Session', sessionSchema);
