--- server/userModel.js
@@ -1,14 +1,0 @@
-//server/userModel.js
-
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-
-const SALT_WORK_FACTOR = 10;
-const bcrypt = require('bcryptjs');
-
-const userSchema = new Schema({
-  username: {type: String, required: true, unique: true},
-  password: {type: String, required: true}
-});
-
-module.exports = mongoose.model('User', userSchema);
